Reuse undo log backing array for entries during transaction release
With the new changes, memtier-benchmark throughput was seen to be low
when run against go-pmem-redis. This is because of the increase in the
number of memory allocations as transaction Release() was allocating a
new backing array for the log entries each time. This commit changes the
abort function to reuse the backing array on transaction Release() and
allocate a new array only in the crash-recovery code path.

Fixes #15.

Signed-off-by: Jerrin Shaji George <jshajigeorge@vmware.com>