Fix case when slice to be logged doesn't exist
If the slice to be logged was never created, i.e data pointer in
sliceheader is nil, we shouldn't create a log entry. If we do create
a log entry, abort() will segfault because we will try to write to
zero address. This commit also adds a test for this case.
The changes are only for undoTx.

Signed-off-by: Mohit Verma <mohitv@vmware.com>