Fix case when slicehdr is in pmem but slice contents are not
Before this commit, if slice header to be logged was in pmem, we
would store the slice contents in log too. But if slice contents were
not in pmem, we would see a segfault on the restart-after-crash path
in *undoTx.abort(). This would happen with the following code:

s := pnew(myStruct)
s.slice = make([]int, 10) // <-- s.slice in pmem but not slice contents
undoTx.Log(&s.slice)
s.slice = pmake([]int, 1)
// CRASH!!!

This commit fixes this case and adds a new test for this case in a
separate directory. We need to actually crash to see this segfault as
only on the restart path, GC zeroes out the volatile pointers stored in
pmem.
The changes are only for undoTx.

Signed-off-by: Mohit Verma <mohitv@vmware.com>