Change how ptr to slice is logged
When ptr to slice is passed to Log() method, user can change the
sliceheader AND/OR the slice contents. So, for undo logging,
1. we store the original sliceheader in the logs.
2. In addition for undo log, we store the original slice contents.
This is useful during transaction abort. When a transaction commits,
we should flush the new sliceheader and the new slice contents. And
we don't need to flush the log entry storing the old slice contents.

For redo logging,
1. We store the new sliceheader in the logs, and when the transaction
commits, we should flush the new sliceheader and new slice contents.

Before this change, we used to have a field in each log entry, indicating
if it stored a sliceheader. This field was stored in pmem and was
unnecessary for all other entries. This commit changes the implementation.
Because special things need to be done only when transaction commits, we
store all this info in volatile memory now.

Testing done: Existing test cases test this feature. All of them pass.

Signed-off-by: Mohit Verma <mohitv@vmware.com>