Change End() to return bool instead of Error
This commit changes End() method to return a bool indicating if
it is safe for the user to release transaction handle or not.
This will also be used by the compiler to figure out when to release
a transaction handle when it is injecting transaction logging statements.
Testing done: Added new tests for undo & redo logging.

Signed-off-by: Mohit Verma <mohitv@vmware.com>