Optimize flush of userdata update at tx End
Previously, If user updates same memory location multiple times
or memory locations in the same cacheline multiple times in the
same transaction, we would issue clwb+fence for same cacheline
multiple times. This commit changes that. Now we track user updates
at the granularity of a cacheline. We now use a RB tree to maintain
addresses, but this can be a map too.
With this change, fannkuch, n-body show perf improvement by ~10%.